requirements.txt
memory-profiler==0.61.0
google-cloud-bigquery==3.11.4
pyspark
regex